+++ lib/ui/web_login_form.dart
@@ -1,0 +1,2 @@
+export 'web_login_form_stub.dart'
+    if (dart.library.js_interop) 'web_login_form_web.dart';